[Rails51] Pass the following tests.
* CoreTest#test_inspect_class
* ReflectionTest#test_association_primary_key_type
* ReflectionTest#test_integer_columns
* ActiveRecord::Migration::ReferencesStatementsTest#test_add_belongs_to_alias
* ActiveRecord::Migration::ReferencesStatementsTest#test_creates_reference_id_column

+++ test/cases/column_test_sqlserver.rb
@@ -36,7 +36,7 @@ class ColumnTestSQLServer < ActiveRecord::TestCase
     it 'bigint(8)' do
       col = column('bigint')
       col.sql_type.must_equal           'bigint(8)'
-      col.type.must_equal               :bigint
+      col.type.must_equal               :integer
       col.null.must_equal               true
       col.default.must_equal            42
       obj.bigint.must_equal             42
